Prevent k from being negative and messing up the hash math

+++ UintahRelease_1_0_0/src/Core/Geom/GeomObj.cc
@@ -56,7 +56,7 @@ MutexPool lock_pool("GeomObj pool", GEOM_LOCK_POOL_SIZE);
 
 static int lock_pool_hash(GeomObj *ptr)
 {
-  long k = ((long)ptr) >> 2; // Disgard unused bits, word aligned pointers.
+  size_t k = ((size_t)ptr) >> 2; // Disgard unused bits, word aligned pointers.
   return (int)((k^(3*GEOM_LOCK_POOL_SIZE+1))%GEOM_LOCK_POOL_SIZE);
 }
